foundry-agenter/orchestrator/requirements.txt
azure-ai-projects>=2.0.0
azure-identity>=1.19.0
azure-data-tables>=12.5.0
aiohttp>=3.10.0
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.0.0
python-dotenv>=1.0.0
# Cache bust: 1775246500
